Use default extension with identity instead of Cassandra specific
