+++ src/k8s-extension/azext_k8s_extension/custom.py
@@ -17,7 +17,7 @@ from ._validators import validate_cc_registration
 
 from .partner_extensions.ContainerInsights import ContainerInsights
 from .partner_extensions.AzureDefender import AzureDefender
-from .partner_extensions.Cassandra import Cassandra
+from .partner_extensions.DefaultExtensionWithIdentity import DefaultExtensionWithIdentity
 from .partner_extensions.AzureMLKubernetes import AzureMLKubernetes
 from .partner_extensions.OpenServiceMesh import OpenServiceMesh
 from .partner_extensions.DefaultExtension import DefaultExtension
@@ -35,7 +35,7 @@ def ExtensionFactory(extension_name):
         'microsoft.azuredefender.kubernetes': AzureDefender,
         'microsoft.openservicemesh': OpenServiceMesh,
         'microsoft.azureml.kubernetes': AzureMLKubernetes,
-        'cassandradatacentersoperator': Cassandra,
+        'cassandradatacentersoperator': DefaultExtensionWithIdentity,
     }
 
     # Return the extension if we find it in the map, else return the default
